+++ src/cmdliner_manpage.mli
@@ -15,6 +15,9 @@ type block =
 type title = string * int * string * string * string
 type t = title * block list
 
+val escape : string -> string
+(** [escape s] escapes [s] from the doc language. *)
+
 (** {1 Standard section names} *)
 
 val s_name : string
@@ -57,13 +60,9 @@ val print :
 
 (** {1 Printers and escapes used by Cmdliner module} *)
 
-val markup_text_escape : string -> string
-(** [markup_text_escape t] escapes [t] to be used as marked up text
-    argument (i.e. doubles ')' or '$'). *)
-
 val subst_vars : Buffer.t -> subst:(string -> string option) -> string -> string
 (** [subst b ~subst s], using [b], substitutes in [s] variables of the form
-    "$(doc)" by their [subst] definition. This leaves $$ escapes and markup
+    "$(doc)" by their [subst] definition. This leaves escapes and markup
     directives $(markup,...) intact.
 
     @raise Invalid_argument in case of illegal syntax. *)